refactor(routing): extract agent module and remove Agent wrapper
